requirements.txt
anyio==4.9.0
APScheduler==3.10.4
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.25.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
nest-asyncio==1.6.0
python-telegram-bot==20.7
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3